Move out input
